Edited make_star_list.c to include (almost) all stars.
Previously there was a magnitude cut off. (Vmag <= 5.5)

The file make_star_list.c is actually compiled as C++ and
contains C++ constructs.  The file is a mess of spaghetti
code.  I want to print a comment when a star is rejected
from inclusion.  Can not do that due to the pastaficated
mess.