Fixed dsc/test/Makefile
The test examples now compile even if the cross compile has
not been run.  So it should be possible to build the test
programs even if a cross compiler is not available.

Restored moon_test.cpp which was lost due to a disk crash.
moon_test.cpp is the test bed for horizontal parallax corrections.